extract interface for comic repository
