+++ androidApp/src/main/java/com/colibrez/xkcdreader/android/ComicViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.savedstate.SavedStateRegistryOwner
-import com.colibrez.xkcdreader.model.Comic
 import com.colibrez.xkcdreader.data.repository.ComicRepository
+import com.colibrez.xkcdreader.model.Comic
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
