+++ shared/src/commonMain/kotlin/com/colibrez/xkcdreader/data/repository/ComicRepository.kt
@@ -1,104 +1,19 @@
 package com.colibrez.xkcdreader.data.repository
 
-import com.colibrez.xkcdreader.model.Comic
-import com.colibrez.xkcdreader.data.model.asExternalModel
 import com.colibrez.xkcdreader.database.model.ComicEntity
-import com.colibrez.xkcdreader.database.model.ComicEntityQueries
-import com.colibrez.xkcdreader.database.model.ComicInfo
-import com.colibrez.xkcdreader.database.model.FavoriteComicEntityQueries
-import com.colibrez.xkcdreader.database.model.ReadComicEntityQueries
-import com.colibrez.xkcdreader.database.model.UserEntity
-import com.colibrez.xkcdreader.database.model.UserEntityQueries
-import com.colibrez.xkcdreader.extensions.getList
-import com.colibrez.xkcdreader.extensions.getOne
-import kotlinx.coroutines.CoroutineDispatcher
+import com.colibrez.xkcdreader.model.Comic
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.withContext
 
-class ComicRepository(
-    val comicQueries: ComicEntityQueries,
-    val readComicQueries: ReadComicEntityQueries,
-    val userEntityQueries: UserEntityQueries,
-    val favoriteComicQueries: FavoriteComicEntityQueries,
-    private val ioDispatcher: CoroutineDispatcher
-) {
+interface ComicRepository {
+    fun getComic(num: Long): Flow<Comic>
+    fun getLatest(): Flow<Comic>
+    fun getCount(): Flow<Long>
+    fun getAllComics(): Flow<List<Comic>>
+    fun getComicsPaged(next: Long, limit: Long): Flow<List<Comic>>
 
-    fun getComic(num: Long): Flow<Comic> {
-        return comicQueries.select(num).getOne(ioDispatcher) { it.asExternalModel()}
-    }
+    suspend fun insertComics(comics: List<ComicEntity>)
 
-    fun getLatest(): Flow<Comic> {
-        return comicQueries.selectLatest().getOne(ioDispatcher) { it.asExternalModel()}
-    }
+    suspend fun markAsSeen(comicNum: Long, userId: Long = 0L)
 
-    fun getCount(): Flow<Long> {
-        return comicQueries.count().getOne(ioDispatcher)
-    }
-
-    fun getAllComics(): Flow<List<Comic>> {
-        return comicQueries.selectAll().getList(ioDispatcher) { it.asExternalModel() }
-    }
-
-    fun getComicsPaged(next: Long, limit: Long): Flow<List<Comic>> {
-        return comicQueries.selectPaged(limit = limit, offset = next).getList(ioDispatcher) { it.asExternalModel() }
-    }
-
-    suspend fun insertComics(comics: List<ComicEntity>) {
-        withContext(ioDispatcher) {
-            comicQueries.transaction {
-                comics.forEach { comic ->
-                    comicQueries.insert(comic)
-                }
-            }
-        }
-    }
-
-    suspend fun markAsSeen(comicNum: Long, userId: Long = 0L) {
-        withContext(ioDispatcher) {
-            userEntityQueries.createUser(id = UserEntity.Id(userId))
-            readComicQueries.markComicAsRead(comicNum, userId = UserEntity.Id(userId))
-        }
-    }
-
-    suspend fun toggleFavorite(comicNum: Long, isFavorite: Boolean, userId: Long = 0L) {
-        withContext(ioDispatcher) {
-            userEntityQueries.createUser(id = UserEntity.Id(userId))
-            if (isFavorite) {
-                favoriteComicQueries.removeComicFromFavorites(comicNumber = comicNum, userId = UserEntity.Id(userId))
-            } else {
-                favoriteComicQueries.markComicAsFavorite(comicNum, userId = UserEntity.Id(userId))
-            }
-        }
-    }
-
-    companion object{
-        fun mapComicSelecting(
-            num: Long,
-            title: String,
-            transcript: String,
-            img: String,
-            alt: String,
-            link: String,
-            year: Long,
-            month: Long,
-            day: Long,
-            isFavorite: Long,
-            isRead: Long,
-        ): Comic {
-            return ComicInfo(
-                num = num,
-                title = title,
-                transcript = transcript,
-                img = img,
-                alt = alt,
-                link = link,
-                year = year,
-                month = month,
-                day = day,
-                isFavorite = isFavorite,
-                isRead = isRead
-            ).asExternalModel()
-        }
-    }
-
+    suspend fun toggleFavorite(comicNum: Long, isFavorite: Boolean, userId: Long = 0L)
 }
